backend/pyproject.toml
[tool.poetry]
name = "package-dashboard"
version = "0.1.0"
description = ""
authors = ["12f23eddde <12f23eddde@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "pkgdash"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
beanie = "^1.18.0"
lxml = "^4.9.2"
tqdm = "^4.65.0"
solv = "^0.7.22.post2"
dynaconf = "^3.1.12"
# python-apt = { url = "https://salsa.debian.org/apt-team/python-apt/-/archive/2.5.3/python-apt-2.5.3.tar.gz" }
aiochclient = "^2.4.0"
loguru = "^0.7.0"
aiohttp = "^3.8.4"
python-dateutil = "^2.8.2"
rocketry = "^2.5.1"
uvicorn = "^0.22.0"
gql = "^3.4.1"
requests-toolbelt = "^1.0.0"
pandas = "^2.0.3"
fastapi-pagination = "^0.12.5"
pygithub = "^2.6.1"
openai = "^1.93.2"
packageurl-python = "^0.17.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
